move schema classes from header to schema.py
